+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `toJson` extension for Any
 - `println` Top-Level Function
+- `asFile` extension for String
+- `ZipFile` Type Class (Experimental API)
 ### Changed
 ### Deprecated
 ### Removed
@@ -155,7 +157,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Overview
 
-* Unreleased - [Diff 1.0.0 to master](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/1.0.0...master)
+* Unreleased - [Diff 1.1.0 to master](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/1.1.0...master)
+* 1.1.0 - [Diff 1.0.0 to 1.1.0](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/1.0.0...1.1.0)
 * 1.0.0 - [Diff 0.2.2 to 1.0.0](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/0.2.2...1.0.0)
 * 0.2.2 - [Diff 0.2.1 to 0.2.2](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/0.2.1...0.2.2)
 * 0.2.1 - [Diff 0.2.0 to 0.2.1](https://gitlab.atvg-studios.at/atvg-studios/kLib/compare/0.2.0...0.2.1)